Not much : - anims creation moved to Preload.create(); - experiments with collision.
Still got a problem with tilemap layer collision.
Got to try a few thing but before i want to get more features working.

@todo : lot of code need cleanup, refactoring.
I'll probably create an entity prefab and use it as a base for player and monster creation.
First i want to have things working and be sure of how things are done.